% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA dataframe}
\usage{
eq_clean_data(quakes_df)
}
\arguments{
\item{quakes_df}{A dataframe with the same column specification as NOAA data}
}
\value{
A modified dataframe containing a 'DATE' column and numeric 'LATITUDE' and 'LONGITUDE' columns
}
\description{
Clean NOAA dataframe
}
\note{
Dates with missing month and day will have these missing values set to 1, and observations with BC dates are filtered.
}
\examples{
\dontrun{
cleaned_df <- eq_clean_data(raw_NOAA_df)
}
}
