% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location column of NOAA earthquake dataset}
\usage{
eq_location_clean(quakes_df)
}
\arguments{
\item{quakes_df}{A dataframe with the same column specification as NOAA earthquake data}
}
\value{
Dataframe containing a LOCATION column with the country stripped out
}
\description{
Clean location column of NOAA earthquake dataset
}
\examples{
\dontrun{
cleaned_df <- eq_location_clean(quakes_df)
}

}
