% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Create map of quakes}
\usage{
eq_map(df, annot_col = "popup_text")
}
\arguments{
\item{df}{A dataframe containing at a minimum, earthquake magnitude, latitude and longitude and a column to label the markers with.}

\item{annot_col}{Name of column containing desired marker label}
}
\value{
Leaflet map
}
\description{
Create map of quakes
}
\examples{
\dontrun{
quakes_df \%>\%
 eq_clean_data() \%>\%
 eq_location_clean() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
}

}
