% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_leaflet.R
\name{eq_map_add_popup}
\alias{eq_map_add_popup}
\title{Add a column with popup text to be displayed on the map}
\usage{
eq_map_add_popup(eq_data)
}
\arguments{
\item{eq_data}{a dataframe with NOAA's earthquake data.}
}
\value{
The original dataframe with a new column called \code{popup_text}.
}
\description{
This function takes the dataset as an argument and creates an
  HTML label that can be used as the annotation text in the leaflet map. This
  function puts together a character string for each earthquake that shows
  the cleaned location (as cleaned by the \code{eq_location_clean()}
  function, the magnitude (\code{EQ_PRIMARY}), and the total number of deaths
  (\code{TOTAL_DEATHS}). If an earthquake is missing values for any of these,
  both the label and the value are skipped for that element of the tag.
}
\examples{
library(dplyr)

# load data
filename <- system.file("extdata/earthquakes.tsv.gz", package = "quakeR")
raw_data <- readr::read_delim(filename, delim = "\\t")

map_data <-
  raw_data \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO",
                lubridate::year(DATE) >= 2000) \%>\%
  eq_map_add_popup()

eq_map(map_data, annot_col = "popup_text")

}
