% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_quaker.R
\name{fit_quaker}
\alias{fit_quaker}
\title{Fit a Logisitc regression to the seismic data and anova analysis of fit}
\usage{
fit_quaker(seismic.geojson.obj)
}
\arguments{
\item{seismic.geojson.obj}{GeoJson object}
}
\value{
Logisitc regression of earthquake data, check if longtitude, latitude, depth, magnitude are good predictors of tsunami occurance.
}
\description{
Fit a Logisitc regression to the seismic data and anova analysis of fit
}
\examples{

data <- fit_quaker(get_seismic_data(timeFrame = 'PAST_WEEK', minMagnitude = '1'))
data
data <- fit_quaker(get_seismic_data(timeFrame = 'PAST_MONTH', minMagnitude = '2.5'))
data
}
\seealso{
\code{\link{get_seismic_data}},
 \code{\link{flatten_to_table}},
 \code{\link{get_freq_grp_by_country_mag}},
 \code{\link{get_freq_grp_by_day_mag}},
 \code{\link{plot}}
}

