% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{geom_timeline_label}
\alias{geom_timeline_label}
\alias{GeomTimelineLabel}
\title{Labels points in a horizontal timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", xmin = NA, xmax = NA, n_max = NA, ...,
  check_overlap = FALSE, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}. If specified and inherit.aes = TRUE (the default), it is
combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:If NULL, the default, the data is inherited from the
plot data as specified in the call to \code{ggplot()}.A data.frame, or other object, will override the plot data. All objects will be
fortified to produce a data frame. See \code{fortify()} for which variables will be created.A function will be called with a
single argument, the plot data. The return value must be a data.frame., and will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function}

\item{xmin}{numeric minimum year to use in the diaplay. If no value for xmin is passed, no filtring is done within the geom for lower
x value dates. If xmin is passed, data rows with x dates below 1 January in xmin year will be filtered out of the display.}

\item{xmax}{numeric maximum year to use in the diaplay. If no value for xmax is passed, no filtring is done within the geom for higher
x value dates. If xmax is passed, data rows with x dates above 31 December in xmax year will be filtered out of the display.}

\item{n_max}{numeric if passed is the largest number of points to be labeled, for example if the y aesthetic identifies three
countries to be included in the chart and n_max has a value of 3, then the 3 largest earthquakes by size for each country will
be labeled. If n_max is not passed then all points are labeled.}

\item{...}{other arguments passed on to \code{layer()}. These are often aesthetics, used to set an aesthetic to a fixed value,
like color = "red" or size = 3. They may also be parameters to the paired geom/stat.}

\item{check_overlap}{logical - TRUE omits text overlapping with other text, FALSE displays all text. Default FALSE}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped.
FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper
functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().}
}
\value{
a ggplot object that can be combined with ggplot or ggmap - this object behaves like a \code{geom_point} ggplot object
}
\description{
\code{geom_timeline_label} is a clone of ggplot2 geom_text that can be used to label points from the NOAA earthquake dataset.
The 'x' aesthetic identify dates to be labeled on the horizontal asix (see supported optional aesthetics below) and the
'optional y' aesthetic can be used to label multiple timelines, for example by selected countries, in a chart.
geom_timeline_label is designed to be an optional addition to geom_timeline.
}
\details{
\code{GeomTimelineLabel} is the \code{ggproto} environment object to enable \code{geom_timeline_label}
}
\section{Aesthetics}{

geom_timeline understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{x}
  \item \strong{label}
}
where x is a vector of type date and label contains the text to be used for the labels

the following aesthetics are optional:
\itemize{
  \item y (can be used as a categorical variable to chart multiple timelines, for example by country)
  \item fontface
  \item lineheight
}
}

\examples{
\dontrun{
my_clean_df \%>\%
  filter(COUNTRY \%in\% c("CHINA", "INDONESIA", "JAPAN")) \%>\%
  filter(lubridate::year(DATE) > 2000) \%>\%
  ggplot(aes(x=DATE, y=COUNTRY, colour=EQ_PRIMARY, alpha=TOTAL_DEATHS,
         size=EQ_PRIMARY, label=LOCATION_NAME)) +
  geom_timeline() +
  geom_timeline_label(n_max = 3) + #label top 3 for each country
  guides(alpha = FALSE, colour=FALSE) +
  scale_x_date(date_labels = "\%Y")
}
}
\references{
for more information on ggplot see http://ggplot2.org/ and http://ggplot2.tidyverse.org/
}
\seealso{
\code{\link{geom_timeline_label}} which can be used to label some of the points
}
\keyword{datasets}
