% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freq_grp_by_country_mag.R
\name{get_freq_grp_by_country_mag}
\alias{get_freq_grp_by_country_mag}
\title{Transforms geojson seismic object to continency table day bay magnitude}
\usage{
get_freq_grp_by_country_mag(seismic.geojson.obj)
}
\arguments{
\item{seismic.geojson.obj}{GeoJson object}
}
\value{
A \code{\link[tibble]{tibble}} with frequency of earthquakes grouped by day and
magnitude
}
\description{
Transforms geojson seismic object to continency table day bay magnitude
}
\examples{
data <- get_freq_grp_by_country_mag(
           get_seismic_data(timeFrame = 'PAST_WEEK', minMagnitude = '1')
        )
data
data <- get_freq_grp_by_country_mag(
           get_seismic_data(timeFrame = 'PAST_MONTH', minMagnitude = '2.5')
        )
data

}
\seealso{
\code{\link{get_seismic_data}},
   \code{\link{fit_quaker}},
   \code{\link{flatten_to_table}},
   \code{\link{get_freq_grp_by_day_mag}},
   \code{\link{plot}}
}

