% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seismic_geojson.R
\name{plot.seismic_geojson}
\alias{plot.seismic_geojson}
\title{Plot geojson object}
\usage{
\method{plot}{seismic_geojson}(x, ...)
}
\arguments{
\item{x}{GeoJson Object}

\item{...}{other arguments to be handle internally by the plot function}
}
\value{
Plots an object of type seismic_geojson.
}
\description{
Plot geojson object
}
\examples{
plot(get_seismic_data(timeFrame = 'PAST_WEEK', minMagnitude = '1'))
plot(get_seismic_data(timeFrame = 'PAST_MONTH', minMagnitude = '2.5'))

}
\seealso{
\code{\link{get_seismic_data}},
 \code{\link{fit_quaker}},
 \code{\link{flatten_to_table}},
 \code{\link{get_freq_grp_by_country_mag}},
 \code{\link{get_freq_grp_by_day_mag}}
}

