% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaker.R
\docType{package}
\name{quaker-package}
\alias{quaker}
\alias{quaker-package}
\title{Clean, plot and map NOAA earthquake data.}
\description{
the \code{quaker} package provides a set of functions to clean and prepare U.S. National Oceanographic and
Atmospheric Administration (NOAA) earthquake data, plot a timeline of earthquakes or map earthquake locations
}
\details{
Functions that might be of interest from \pkg{quaker} include\cr
\code{\link{eq_clean_data}} to clean and prepare the NOAA data after it has been read from the website\cr
\code{\link{geom_timeline}} to plot earthquake activity over time\cr
\code{\link{geom_timeline_label}} to annotate a geom_timeline plot with labels\cr
\code{\link{eq_map}} to map earthquake locations\cr
Refer to the vignette for more information.
}
