% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaker_shiny.R
\name{quaker_shiny}
\alias{quaker_shiny}
\title{Shiny app forthe quaker package}
\usage{
quaker_shiny()
}
\value{
Runs the shiny app
}
\description{
Shiny app forthe quaker package
}
\examples{
\dontrun{
quaker_shiny()
}
}

