% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quakes}
\alias{quakes}
\title{Example NOAA earthquake data}
\format{A table data frame with 64 rows of 47 variables (see the data definition reference below for full description):
\describe{
  \item{I_D}{integer row ID}
  \item{FLAG_TSNAMI}{character - contains "Tsu" where these was a tsunami event}
  \item{YEAR}{integer year of the quake - can be negative to denote B.C.E year}
  ...
}}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
and the data defintions can be found at https://www.ngdc.noaa.gov/nndc/struts/results?&t=101650&s=225&d=225
}
\usage{
quakes
}
\description{
A dataset containing 64 rows of the NOAA earthquake dataset to provide sample data for the quaker routines
}
\keyword{datasets}
