% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA dataset}
\usage{
eq_clean_data(raw)
}
\arguments{
\item{raw}{raw NOAA data frame}
}
\value{
a clean data frame
}
\description{
Takes raw NOAA data frame and returns a clean data frame. The clean data
frame has the following: A date column created by uniting the year,
month, day and converting it to the Date class LATITUDE and LONGITUDE columns
converted to numeric class.
}
\examples{
\dontrun{
data <- readr::read_delim("earthquakes.tsv.gz", delim = "\\t")
clean_data <- eq_clean_data(data)
}

}
