% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the LOCATION_NAME column of NOAA dataset}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{character vector from LOCATION_NAME column of NOAA dataset}
}
\value{
character vector of clean location names
}
\description{
This function cleans the LOCATION_NAME column by stripping out the country
name (including the colon) and converts names to title case (as opposed to
all caps).
}
\examples{
\dontrun{
data <- readr::read_delim("earthquakes.tsv.gz", delim = "\\t")
clean_data <- eq_location_clean(data)
}

}
\keyword{internal}
