% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Plot earthquakes on an interactive map}
\usage{
eq_map(df, annot_col = "DATE")
}
\arguments{
\item{df}{a data containing the filtered data frame with earthquakes to
visualize.}

\item{annot_col}{the name of the column used for the annotation in
the pop-up Defaults to "DATE".}
}
\value{
a \code{leaflet} type HTML widget.
}
\description{
The function maps the epicenters (LATITUDE/LONGITUDE) and
  annotates each point with a pop up window containing annotation data stored
  in a column of the data frame. Each earthquake is shown with a circle, and
  the radius of the circle is proportional to the earthquake’s magnitude
  (EQ_PRIMARY).
}
\examples{
library(dplyr)

# load data
filename <- system.file("extdata/earthquakes.tsv.gz", package = "quakeR")
raw_data <- readr::read_delim(filename, delim = "\\t")

map_data <-
  raw_data \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO",
                lubridate::year(DATE) >= 2000)

eq_map(map_data)
eq_map(map_data, annot_col = "EQ_PRIMARY")
eq_map(map_data, annot_col = "LOCATION_NAME")

}
