% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Map NOAA earthquake data}
\usage{
eq_map(df, annot_col, pretty = FALSE)
}
\arguments{
\item{df}{is the raw NOAA data frame as read from the web site - a link to the data source and to the data descriptions
can be found below in the refereces}

\item{annot_col}{is a single text value for the column name to be displayed as the popup value when the user hovers over
a data point on the map. This column can be a single data item, for example "DATE" or this can be a column of combined
values using the \code{eq_create_label} function.}

\item{pretty}{single logical value. If pretty is true, the columne name of 'annot_col' is displayed prior to the value
in the popup. If pretty is FALSE or not passed, only the value is displayed.}
}
\value{
the cleaned NOAA data frame
}
\description{
\code{eq_map} displays an interactive map showing earthquake locations from the NOAA earthquake data. \code{eq_map} uses
the leaflet R package to
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
 eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "DATE", pretty = TRUE)
}
}
\references{
The NOAA earthquake database https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1 and the data defintions
can be found at https://www.ngdc.noaa.gov/nndc/struts/results?&t=101650&s=225&d=225
}
\seealso{
\code{\link{eq_create_label}} to create popup display with multiple columns and
\code{\link{eq_clean_data}} to prepare the NOAA data
}
